A test of login, and som rearrangements to accomodate it
